--- app/routes/items/search_code.js
@@ -1,24 +1,0 @@
-import Ember from 'ember';
-import AuthorizeRoute from './../authorize';
-
-export default AuthorizeRoute.extend({
-
-  model(params) {
-    var item = this.store.peekRecord("item", params.item_id);
-    return Ember.RSVP.hash({
-      item: item || this.store.findRecord('item', params.item_id),
-      codes: this.store.query('code', { stock: true })
-    });
-  },
-
-  setupController(controller, model){
-    this._super(controller, model);
-    controller.set('item', model.item);
-    controller.set('searchText', "");
-  },
-
-  renderTemplate: function(controller) {
-    this.render('search_code', {controller: controller});
-  }
-
-});
